requirements.txt
python-dotenv==1.0.0
dashscope==1.14.0
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.2.2
chromadb==0.4.24
hnswlib==0.8.0
rank-bm25==0.2.2
jieba==0.42.1
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4